# api/v2/openapi.yml
openapi: "3.0.2"
info:
  title: Authcore API
  version: "2.0"
paths:
...
  /api/v2/authn:
    post:
      summary: Start an local authentication transaction
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                handle:
                  type: string
                code_challenge_method:
                  type: string
                code_challenge:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
              required:
                - client_id
                - handle
                - redirect_uri
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/password:
    post:
      summary: Request a password key exchange challenge
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state_token:
                  type: string
                message:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
  /api/v2/authn/password/verify:
    post:
      summary: Verify password factor
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state_token:
                  type: string
                verifier:
                  type: string
              required:
                - state_token
                - verifier
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/mfa/{method}:
    post:
      summary: Request a MFA challenge
      tags:
        - authn
      parameters:
        - in: path
          name: method
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                state_token:
                  type: string
                message:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
  /api/v2/authn/mfa/{method}/verify:
    post:
      summary: Verify a MFA factor
      tags:
        - authn
      parameters:
        - in: path
          name: method
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                state_token:
                  type: string
                verifier:
                  type: string
              required:
                - state_token
                - verifier
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/idp/{provider}:
    post:
      summary: Start a third-party IDP authentication transaction
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
              required:
                - client_id
                - handle
                - redirect_uri
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/idp/{provider}/verify:
    post:
      summary: Verify a third-party IDP authorization grant
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state_token:
                  type: string
                code:
                  type: string
              required:
                - state_token
                - code
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/idp_binding/{provider}:
    post:
      summary: Start a third-party IDP binding transaction
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
  /api/v2/authn/idp_binding/{provider}/verify:
    post:
      summary: Verify a third-party IDP authorization grant for a third-party IDP binding transaction.
      tags:
        - authn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state_token:
                  type: string
                code:
                  type: string
              required:
                - state_token
                - code
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
...
  /api/v2/signup:
    post:
      summary: Create a new user using primary authentication method.
      tags:
        - signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                password_verifier:
                  $ref: '#/components/schemas/PasswordVerifier'
                redirect_uri:
                  type: string
                  format: uri
                email:
                  type: string
                phone:
                  type: string
                name:
                  type: string
              required:
                - client_id
                - password_verifier
                - redirect_uri
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthnState"
...
components:
  schemas:
    AuthnState:
      type: object
      properties:
        state_token:
          type: string
        status:
          type: string
        password_method:
          type: string
        password_salt:
          type: string
        factors:
          type: array
          items:
            type: string
        idp_authorization_url:
          type: string
          format: uri
        authorization_code:
          type: string
      required:
        - state_token
        - status
    PasswordVerifier:
      oneOf:
        - $ref: '#/components/schemas/SPAKE2PlusVerifier'
    SPAKE2PlusVerifier:
      type: object
      properties:
        method:
          type: string
        salt:
          type: string
        w0:
          type: string
        l:
          type: string
      required:
        - method
        - salt
        - w0
        - l
...